+++ packages/common/src/initiatives/_internal/InitiativeUiSchemaEdit.ts
@@ -7,6 +7,7 @@ import { getThumbnailUiSchemaElement } from "../../core/schemas/internal/getThum
 import { IHubInitiative } from "../../core";
 import { getAuthedImageUrl } from "../../core/_internal/getAuthedImageUrl";
 import { fetchCategoriesUiSchemaElement } from "../../core/schemas/internal/fetchCategoriesUiSchemaElement";
+import { getSlugSchemaElement } from "../../core/schemas/internal/getSlugSchemaElement";
 
 /**
  * @private
@@ -181,6 +182,7 @@ export const buildUiSchema = async (
         type: "Section",
         labelKey: `${i18nScope}.sections.searchDiscoverability.label`,
         elements: [
+          getSlugSchemaElement(i18nScope),
           ...(await fetchCategoriesUiSchemaElement(i18nScope, context)),
           {
             labelKey: `${i18nScope}.fields.tags.label`,