+++ packages/common/test/discussions/_internal/DiscussionUiSchemaEdit.test.ts
@@ -5,9 +5,14 @@ import * as getLocationExtentModule from "../../../src/core/schemas/internal/get
 import * as getLocationOptionsModule from "../../../src/core/schemas/internal/getLocationOptions";
 import * as getTagItemsModule from "../../../src/core/schemas/internal/getTagItems";
 import { UiSchemaRuleEffects } from "../../../src/core/schemas/types";
+import * as getSlugSchemaElementModule from "../../../src/core/schemas/internal/getSlugSchemaElement";
 
 describe("buildUiSchema: discussion edit", () => {
-  it("returns the full discussion edit uiSchema", async () => {
+  const mockSlugElement = {
+    labelKey: "slug",
+    scope: "/properties/_slug",
+  } as any;
+  beforeEach(() => {
     spyOn(fetchCategoryItemsModule, "fetchCategoryItems").and.returnValue(
       Promise.resolve([
         {
@@ -25,7 +30,11 @@ describe("buildUiSchema: discussion edit", () => {
     spyOn(getTagItemsModule, "getTagItems").and.returnValue(
       Promise.resolve([])
     );
-
+    spyOn(getSlugSchemaElementModule, "getSlugSchemaElement").and.returnValue(
+      mockSlugElement
+    );
+  });
+  it("returns the full discussion edit uiSchema", async () => {
     const uiSchema = await buildUiSchema(
       "some.scope",
       {
@@ -151,6 +160,7 @@ describe("buildUiSchema: discussion edit", () => {
           type: "Section",
           labelKey: "some.scope.sections.searchDiscoverability.label",
           elements: [
+            mockSlugElement,
             {
               labelKey: "some.scope.fields.tags.label",
               scope: "/properties/tags",
@@ -261,24 +271,6 @@ describe("buildUiSchema: discussion edit", () => {
     });
   });
   it("returns the full discussion edit uiSchema when entity does have a view", async () => {
-    spyOn(fetchCategoryItemsModule, "fetchCategoryItems").and.returnValue(
-      Promise.resolve([
-        {
-          value: "/categories",
-          label: "/categories",
-        },
-      ])
-    );
-    spyOn(getLocationExtentModule, "getLocationExtent").and.returnValue(
-      Promise.resolve([])
-    );
-    spyOn(getLocationOptionsModule, "getLocationOptions").and.returnValue(
-      Promise.resolve([])
-    );
-    spyOn(getTagItemsModule, "getTagItems").and.returnValue(
-      Promise.resolve([])
-    );
-
     const uiSchema = await buildUiSchema(
       "some.scope",
       {
@@ -407,6 +399,7 @@ describe("buildUiSchema: discussion edit", () => {
           type: "Section",
           labelKey: "some.scope.sections.searchDiscoverability.label",
           elements: [
+            mockSlugElement,
             {
               labelKey: "some.scope.fields.tags.label",
               scope: "/properties/tags",