+++ packages/common/CHANGELOG.md
@@ -1,3 +1,24 @@
+# @esri/hub-common [14.223.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.222.0...@esri/hub-common@14.223.0) (2024-10-30)
+
+
+### Features
+
+* **hub-common:** add slug to entity edit schemas ([#1706](https://github.com/Esri/hub.js/issues/1706)) ([dc38ff4](https://github.com/Esri/hub.js/commit/dc38ff4fe48386ce1fec59a1f05a14ec23690e62))
+
+# @esri/hub-common [14.222.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.221.1...@esri/hub-common@14.222.0) (2024-10-29)
+
+
+### Features
+
+* **hub-common:** add channelAcl to PostRelation ([#1701](https://github.com/Esri/hub.js/issues/1701)) ([5568041](https://github.com/Esri/hub.js/commit/5568041765d4672deb486b485a463c926d7f8257))
+
+## @esri/hub-common [14.221.1](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.221.0...@esri/hub-common@14.221.1) (2024-10-29)
+
+
+### Bug Fixes
+
+* **hub-common:** remove calcite-radio-group ([#1704](https://github.com/Esri/hub.js/issues/1704)) ([1daf1bc](https://github.com/Esri/hub.js/commit/1daf1bc756a4a47067237913e2e65bed4497d00f))
+
 # @esri/hub-common [14.221.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.220.0...@esri/hub-common@14.221.0) (2024-10-25)